+++ CHANGELOG.md
@@ -12,9 +12,11 @@
 - Fix fastapi deprecated events
 - Easier setting and getting metadata values
 
-## [0.0.9] - 08/12/2025
+## [1.0.0] - 25/12/2025
 
 * Add pluggable data_store module with in-memory and Postgres backends
+* Add in memory event queue
+* add adapters(OpenAI, Gemini, Anthropic) for support streaming event
 
 ## [0.0.8] - 22/04/2025